requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
tenacity>=8.3.0
orjson>=3.10.0
python-dotenv>=1.0.1

# LangChain + providers
langchain>=0.2.10
langchain-openai>=0.1.18
# Uncomment if/when needed
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.21
langchain-community>=0.2.0
